os/bluestore: allow latency to adpatively manage the size of the max bytes throttle
Uses the commit latency of the transactions to adapt the max bytes
throttle setting. The maximum number of bytes will be set to a power
of the square root of 2. When the latency is above the maximum latency
setting, the throttle size is increased by one step (one power of the
square root of 2). When the latency is below the minimum latency
setting the throttle size is decrease by one step. The other throttle
(throttle_deferred_bytes) is not used when latency managed throttle is
active.

When latency managed throttle is engaged, the max ops, deferred
max ops, and deferred max bytes throttles are not used.

Adds four new configuration options (the last 3 can be changed while
running):

  * bluestore_latency_managed_throttle     : whether to use the managed throttle
  * bluestore_throttle_latency_min         : minimum latency
  * bluestore_throttle_latency_max         : maximum latency
  * bluestore_throttle_latency_window_msec : duration of window in which
                                             average latencies are computed in
                                             milliseconds

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>